+++ packages/adminui-vue/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.5]
+### Changed
+ - !!Breaking!! `u-auto-field` removed v-model, now gets and sets value from store automatically
+ - !!Breaking!! `u-auto-field` renamed prop `code` to `attribute-name`
+
 ## [1.6.4]
 ### Fixed
  - "Always use domain for login" checkbox should be visible in navbar User menu in case "Negotiate" auth is enabled for app [UBDF-9586]